Removed TrustedHttps (todo) Bumped dependencies. Ditched 2.9.x, published for 2.10.4 & 2.11.2

+++ util/src/main/scala/hr/element/etb/Runner.scala
@@ -7,11 +7,11 @@ package object io {
   import java.io.{ File, ByteArrayInputStream, ByteArrayOutputStream }
   import java.lang.{ Thread, Runnable, InterruptedException }
 
-  implicit val WaitPeriod = 60000
+  implicit val WaitPeriod = 60000L
   implicit val WorkingDir = new File(".")
 
   object Runner {
-    def apply(cmd: Seq[String], input: Array[Byte] = Array())(implicit workingDir: File, waitPeriod: Int) = {
+    def apply(cmd: Seq[String], input: Array[Byte] = Array())(implicit workingDir: File, waitPeriod: Long) = {
       val pb = Process(cmd, workingDir)
 
       val oS = new ByteArrayOutputStream
